requirements.txt
streamlit
pandas
numpy
plotly
yfinance
requests
reportlab
scipy
python-dotenv